[UPD] feat(mails): New cc for dissertation mails.
Updated the recipient logic:
- Deposit emails are sent to authors and a specfic manager adress.
- Archive emails are sent to authors, promoters and a specific manager address.

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>
